Creation of Dice Score folder
assigning every Dice Score material to one Folder.
