+++ Dice_Score/Dice Score main.py
@@ -1,0 +1,44 @@
+from skimage.io import imread
+from skimage.filters import threshold_otsu  # otsu-global pakage
+
+
+# skimage Otsu-Global
+image = imread("data-git/N2DH-GOWT1/img/t01.tif", as_gray=True)
+otsu_threshold = threshold_otsu(image)
+otsu = (image > otsu_threshold).astype(int).flatten()  # binary & 1D
+
+# start of Dice Score code
+from skimage.io import imread
+
+
+otsu_img = otsu   # output of otsu
+
+ground_truth = imread("data-git/N2DH-GOWT1/gt/man_seg01.tif", as_gray=True)
+otsu_gt = (ground_truth > 0).astype(int).flatten()  # gt binary & 1D
+
+
+def dice_score(otsu_img, otsu_gt):
+
+    # control if the Pictures have the same Size
+    if len(otsu_img) != len(otsu_gt):
+        raise ValueError("Images don't have the same length!")
+
+    # defining the variables for the Dice Score equation
+    positive_overlap = 0
+    sum_img = 0
+    sum_gt = 0
+
+    for t, p in zip(otsu_img, otsu_gt):
+        if t == 1:
+            sum_img += 1
+        if p == 1:
+            sum_gt += 1
+        if t == 1 and p == 1:
+            positive_overlap += 1
+
+    if sum_img + sum_gt == 0:
+        return 1.0
+
+    return 2 * positive_overlap / (sum_img + sum_gt)
+
+print("Dice Score:", dice_score(otsu_img, otsu_gt))
